Add publicapi to cache instead of frontend-lb
In an attempt to completely remove frontend-lb, add publicapi
application to cache instances instead.

The plan is to update the DNS record to point at the cache elb, so it
will automatically forward requests to itself, and then nginx will pick
up requests with the host header and forward requests to the relevant
machines.